templates/python-basic/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional dependencies (uncomment as needed)
# pandas>=2.0.0
# numpy>=1.24.0
# flask>=3.0.0
# fastapi>=0.100.0
# uvicorn>=0.23.0